# QSDM/docs/docs/openapi.yaml
openapi: 3.0.3
info:
  title: QSDM HTTP API
  description: |
    Quantum-Secure Dynamic Mesh Ledger (QSDM, formerly QSDM) HTTP API with
    military-grade security. Native coin: Cell (CELL), 8 decimals, smallest
    unit `dust`.
    
    All authenticated endpoints require:
    - JWT Bearer token in Authorization header
    - Request signing for POST/PUT/DELETE requests (X-Timestamp, X-Nonce, X-Signature headers)
    
    Security Features:
    - TLS 1.3 encryption
    - Quantum-safe JWT tokens (CRYSTALS-Dilithium)
    - Request signing with quantum-safe signatures
    - Rate limiting: default **100 requests per client per minute** (`[api] rate_limit_max_requests` / `rate_limit_window` or `QSDM_API_RATE_LIMIT_MAX` / `QSDM_API_RATE_LIMIT_WINDOW`; the legacy `QSDM_*` env vars continue to be accepted during the deprecation window, see `REBRAND_NOTES.md`). Some routes are stricter (e.g. GET /monitoring/ngc-challenge at 15/min). Health checks and public trust/audit transparency reads are exempt so probes and public widgets are not throttled.
    - Replay attack protection
    - Role-based access control
    
...

...
  version: 1.1.0
  contact:
    name: QSDM Development Team
    url: https://github.com/blackbeardONE/QSDM
servers:
  - url: https://localhost:8443/api/v1
    description: Production server (HTTPS)
  - url: http://localhost:8080/api/v1
    description: Development server (HTTP - insecure)
...
paths:
...
  /monitoring/ngc-challenge:
    get:
      tags:
        - Monitoring
      summary: Issue ingest nonce (NVIDIA-lock nonce mode)
      description: |
        Public for JWT purposes; requires `X-QSDM-NGC-Secret` (or legacy `X-QSDM-NGC-Secret`) matching the node's ingest secret.
        Returns 404 when `nvidia_lock_require_ingest_nonce` is disabled. Rate limited to **15 requests/minute** per client (IP or X-API-Key).
        On 429, `Retry-After` is set (seconds).
      operationId: ngcIngestChallenge
      parameters:
        - $ref: '#/components/parameters/NgcIngestSecretPrimary'
        - $ref: '#/components/parameters/NgcIngestSecretLegacy'
      responses:
        '200':
          description: Nonce issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  qsdm_ingest_nonce:
                    type: string
                  expires_at_unix:
                    type: integer
                    format: int64
                  ttl_seconds:
                    type: integer
        '404':
          description: Ingest nonce challenge disabled
        '401':
          description: Missing or invalid ingest secret
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
                description: Seconds until the client may retry
  /monitoring/ngc-proof:
    post:
      tags:
        - Monitoring
      summary: Ingest NGC GPU proof bundle
      description: |
        Accepts JSON proof bundles from the NGC sidecar. Requires ingest secret headers.
        Rate limited (default 30/min per path key). See deploy README for NVIDIA-lock and HMAC/nonce behavior.
      operationId: ngcProofIngest
      parameters:
        - $ref: '#/components/parameters/NgcIngestSecretPrimary'
        - $ref: '#/components/parameters/NgcIngestSecretLegacy'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Proof bundle (cuda_proof_hash, gpu_fingerprint, optional qsdm_* fields)
      responses:
        '200':
          description: Ingest accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '400':
          description: Invalid body or proof validation failed
        '404':
          description: NGC ingest not configured on node
        '413':
          description: Body too large
        '429':
          description: Rate limit exceeded
  /monitoring/ngc-proofs:
    get:
      tags:
        - Monitoring
      summary: List ingested proof summaries
      description: Operator view of recent ingested bundles (same secret as ingest).
      operationId: ngcProofList
      parameters:
        - $ref: '#/components/parameters/NgcIngestSecretPrimary'
        - $ref: '#/components/parameters/NgcIngestSecretLegacy'
      responses:
        '200':
          description: Summary list
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  proofs:
                    type: array
                    items:
                      type: object
        '404':
          description: NGC ingest not configured
        '429':
          description: Rate limit exceeded
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Register a new user account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - address
                - password
              properties:
                address:
                  type: string
                  description: Wallet address
                  example: "wallet_address_123"
                password:
                  type: string
                  format: password
                  minLength: 8
                  description: User password (minimum 8 characters)
                  example: "<your-strong-password>"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "user registered successfully"
                  address:
                    type: string
        '400':
          description: Invalid input
        '409':
          description: User already exists
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: Login and receive JWT tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - address
                - password
              properties:
                address:
                  type: string
                  example: "wallet_address_123"
                password:
                  type: string
                  format: password
                  example: "<your-strong-password>"
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token (15 minutes)
                  refresh_token:
                    type: string
                    description: JWT refresh token (7 days)
                  expires_in:
                    type: integer
                    description: Access token expiration in seconds
                    example: 900
        '401':
          description: Invalid credentials
...
  /wallet/create:
    post:
      tags:
        - Wallet
      summary: Create wallet
      description: Create a new wallet (requires authentication)
      operationId: createWallet
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Wallet created
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    type: string
                  balance:
                    type: number
                    format: float
        '401':
          description: Unauthorized
        '500':
          description: Failed to create wallet (e.g. crypto or key generation error; check server logs)
  /wallet/balance:
    get:
      tags:
        - Wallet
      summary: Get wallet balance
      description: Get balance for a wallet address
      operationId: getBalance
      security:
        - bearerAuth: []
      parameters:
        - name: address
          in: query
          description: Wallet address (defaults to authenticated user's address)
          schema:
            type: string
      responses:
        '200':
          description: Balance retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    type: string
                  balance:
                    type: number
                    format: float
        '400':
          description: Missing address (query and no authenticated user address)
        '401':
          description: Unauthorized
        '500':
          description: Failed to read balance from storage
...
  /wallet/mint:
    post:
      tags:
        - Wallet
      summary: '[REMOVED in v0.3.3] Mint main coin ($CELL) — now returns 410 Gone'
      deprecated: true
      description: |
        **Removed in v0.3.3 (Session 91).** This endpoint was a non-functional
        stub in v0.3.2 and earlier: it accepted `{recipient, amount}`, returned
        HTTP 200 with `status:"minted"`, but **never credited the recipient's
        balance** — there was no code path from this handler to the wallet
        service's `AddBalance` operation. A subsequent
        `GET /wallet/balance?address=<recipient>` always returned 0.

        The handler now returns **HTTP 410 Gone** with a JSON body carrying a
        structured `migration` block. Callers should switch to:

        * **New-operator funding:** `POST /wallet/send` from an existing
          operator (peer transfer; see `MINER_QUICKSTART.md` Appendix B), or
          wait for the public incentivised testnet (tracked as `mining-05`
          in `RELEASE_NOTES_v0.3.0.md`).
        * **Named token mint:** `POST /tokens/mint` is the working,
          wallet-service-backed mint path for secondary tokens.

        The `qsdm_wallet_mint_total{result="gone"}` Prometheus counter
        increments for every call to this path on v0.3.3+ nodes — a rising
        value indicates a misconfigured caller that should be migrated.
      operationId: mintMainCoin
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              description: 'Body is ignored — handler always returns 410 Gone.'
      responses:
        '410':
          description: Endpoint removed in v0.3.3 — see `migration` field for next steps.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: gone
                  reason:
                    type: string
                  migration:
                    type: object
                    properties:
                      new_operator_funding:
                        type: string
                      named_token_minting:
                        type: string
                      public_testnet:
                        type: string
        '405':
          description: Method not allowed (only POST is routed; GET/PUT/DELETE return 405 before the 410 evaluation).
...
  /wallet/nonce:
    get:
      tags:
        - Wallet
      summary: Read next nonce for a self-custody envelope
      description: |
        v0.4.1 (Session 100) public-read helper that lets a self-custody
        client fetch the next envelope nonce without an authenticated
        session. Symmetric with `/wallet/balance`: read-only,
        address-required-in-query, no JWT. The returned `nonce` is the
        next allowed value the chain will accept for a `tx.nonce` field
        on this address. Replay protection is enforced server-side by
        `pkg/storage` on `submit-signed`. See
        `V041_REPLAY_PROTECTION_DESIGN.md` §5.2.
      operationId: walletNonce
      parameters:
        - name: address
          in: query
          required: true
          schema:
            type: string
          description: Wallet address (sender = hex(sha256(public_key)))
      responses:
        '200':
          description: Next allowed nonce
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    type: string
                  nonce:
                    type: integer
                    format: int64
                    description: Next acceptable transaction nonce
        '400':
          description: Missing or malformed `address`

  # --------------------------------------------------------------------------
  # Audit — transparency surface (Session 77 / v0.4.2)
  # --------------------------------------------------------------------------
  /audit/summary:
    get:
      tags:
        - Audit
      summary: Audit checklist summary
      description: |
        Public-API mirror of the dashboard's `/api/audit/summary`
        (bearer-gated). Returns bucket counts, score, blocking-findings
        flag and preview, and a 4-bucket evidence-provenance breakdown
        of the runtime-verified items in `pkg/audit/checklist.go`.

        Wire shape is byte-for-byte identical to the dashboard view, so
        an SDK client can drop-replace
        `https://dashboard.qsdm.tech/api/audit/summary` with
        `https://api.qsdm.tech/api/v1/audit/summary` and get the same
        JSON. Pinned by `TestAuditAPI_WireParity_DashboardAndAPI`.

        Cache: `Cache-Control: public, max-age=60`. A flip is a Git
        commit + redeploy event; 60 s of staleness is acceptable.
      operationId: auditSummary
      responses:
        '200':
          description: Current audit checklist summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditSummary'
        '405':
          description: Method not allowed (only GET)
  /audit/items:
    get:
      tags:
        - Audit
      summary: Audit checklist items (filterable)
      description: |
        Full filterable items list, public mirror of the dashboard's
        `/api/audit/items`. All three filters are closed-enum
        validated against `pkg/audit` constants — a typo'd value
        returns **400**, NOT a silent "all items" passthrough.
        Applied filters are echoed back via `filters` (omitted when
        no filter is set).
      operationId: auditItems
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum:
              - cryptography
              - authentication
              - authorisation
              - network
              - smart_contracts
              - bridge
              - storage
              - api
              - governance
              - infrastructure
              - supply_chain
              - runtime
              - secret_rotation
              - rebrand
              - tokenomics
              - mining_audit
              - trust_api
        - name: severity
          in: query
          required: false
          schema:
            type: string
            enum: [critical, high, medium, low, info]
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, passed, failed, waived]
      responses:
        '200':
          description: Filtered items list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditItemsResponse'
        '400':
          description: Closed-enum filter parameter mis-typed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method not allowed
...
  /streams:
    get:
      tags:
        - Streams
      summary: List CELL streams
      operationId: listStreams
      parameters:
        - name: payer
          in: query
          schema:
            type: string
        - name: provider
          in: query
          schema:
            type: string
        - name: service_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, paused, closed]
      responses:
        '200':
          description: Consensus stream projection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamsResponse'
        '503':
          description: Stream state is not configured
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token signed with CRYSTALS-Dilithium (quantum-safe).
        Token format: base64(claims) + "." + base64(signature)
        
        For POST/PUT/DELETE requests, also include:
        - X-Timestamp: Unix timestamp (seconds)
        - X-Nonce: Unique nonce (32 bytes, base64)
        - X-Signature: Request signature (base64, quantum-safe)
...